requirements.txt
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info===1.1build1
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
fpdf==1.7.2
fpdf2==2.7.9
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
pillow==10.4.0
platformdirs==4.2.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
python-apt==2.4.0+ubuntu1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.26.3
wadllib==1.3.6
zipp==1.0.0
